fix: dock Every bottom sheet to bottom of screen
